feat: added BatteryIndicator

+++ composeApp/src/commonMain/kotlin/hangar/domain/DroneStatus.kt
@@ -3,4 +3,4 @@ package hangar.domain
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class DroneStatus(val state: DroneState, val battery: Int?, val location: String?)
+data class DroneStatus(val state: DroneState, val battery: Float?, val location: String?)
